Organize routes into multiple files according to auth scope
